// src/api-client/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EntityBasic } from './entity-basic';

/**
 * 
 * @export
 * @interface UserResponse
 */
export interface UserResponse {
    /**
     * 
     * @type {number}
     * @memberof UserResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'status'?: UserResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'avatarUrl'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'dob'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'gender'?: UserResponseGenderEnum;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'updatedBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {Set<EntityBasic>}
     * @memberof UserResponse
     */
    'roles'?: Set<EntityBasic>;
}

export const UserResponseStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;

export type UserResponseStatusEnum = typeof UserResponseStatusEnum[keyof typeof UserResponseStatusEnum];
export const UserResponseGenderEnum = {
    Female: 'FEMALE',
    Male: 'MALE',
    Other: 'OTHER'
} as const;

export type UserResponseGenderEnum = typeof UserResponseGenderEnum[keyof typeof UserResponseGenderEnum];