// src/api-client/models/role-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RoleResponse
 */
export interface RoleResponse {
    /**
     * 
     * @type {number}
     * @memberof RoleResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof RoleResponse
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleResponse
     */
    'description'?: string;
}